break apart report and data management ui
