+++ ui/pages/DataManagement/components/SampleQc.jsx
