+++ ui/pages/DataManagement/selectors.js
@@ -1,0 +1,5 @@
+export const getElasticsearchStatusLoading = state => state.elasticsearchStatusLoading.isLoading
+export const getElasticsearchStatusData = state => state.elasticsearchStatus
+export const getQcUploadStats = state => state.qcUploadStats
+export const getAllUsersLoading = state => state.allUsersLoading.isLoading
+export const getAllUsers = state => state.allUsers
