+++ ui/pages/Report/components/SeqrStats.jsx
