+++ ui/pages/Report/constants.js
@@ -2,7 +2,6 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 
 import PedigreeImagePanel from 'shared/components/panel/view-pedigree-image/PedigreeImagePanel'
-import { successStoryTypeDisplay } from 'shared/utils/constants'
 
 export const STAFF_SEARCH_FORM_NAME = 'customVariantSearch'
 export const INCLUDE_ALL_PROJECTS = 'allProjectFamilies'
@@ -82,15 +81,6 @@ const formatFamilySummary = row =>
     }
   </div>
 
-const formatIDLink = row =>
-  <Link to={`/project/${row.project_guid}/family_page/${row.family_guid}`} target="_blank">{row.family_id}</Link>
-
-const formatSuccessStoryTypes = row =>
-  row.success_story_types && row.success_story_types.map(tag => <div>{successStoryTypeDisplay(tag)}</div>)
-
-const formatDiscoveryTags = row =>
-  row.extras_variant_tag_list && <div>{row.extras_variant_tag_list.map(tag => <div><small>{tag}</small></div>)}</div>
-
 export const DISCOVERY_SHEET_COLUMNS = [
   { name: 't0', content: 'T0', format: formatT0, style: { minWidth: '100px' } },
   { name: 'family_id', content: 'Family ID', format: formatFamilySummary, noFormatExport: true, style: { minWidth: '200px' } },
@@ -152,11 +142,4 @@ export const DISCOVERY_SHEET_COLUMNS = [
   { name: 'collaborator', content: 'Collaborator' },
   { name: 'num_individuals_sequenced', content: '# of Individuals Sequenced' },
   { name: 'analysis_summary', content: 'Analysis Summary', style: { minWidth: '800px' } },
-]
-
-export const SUCCESS_STORY_COLUMNS = [
-  { name: 'family_id', content: 'Family ID', format: formatIDLink, noFormatExport: true, style: { minWidth: '200px' } },
-  { name: 'success_story_types', content: 'Success Story Types', format: formatSuccessStoryTypes, noFormatExport: true, style: { minWidth: '300px' } },
-  { name: 'success_story', content: 'Success Story', style: { minWidth: '564px' } },
-  { name: 'discovery_tags', content: 'Discovery Tags', format: formatDiscoveryTags, noFormatExport: true, style: { minWidth: '400px' } },
 ]
