+++ ui/pages/Report/reducers.js
@@ -1,63 +1,21 @@
 import { combineReducers } from 'redux'
-import { SubmissionError } from 'redux-form'
 
-import { loadingReducer, createSingleValueReducer, createSingleObjectReducer } from 'redux/utils/reducerFactories'
-import { RECEIVE_DATA, REQUEST_SAVED_VARIANTS, REQUEST_PROJECTS, loadProject } from 'redux/rootReducer'
-import { SHOW_ALL, SORT_BY_XPOS } from 'shared/utils/constants'
+import { loadingReducer, createSingleValueReducer } from 'redux/utils/reducerFactories'
+import { RECEIVE_DATA, loadProject } from 'redux/rootReducer'
 import { HttpRequestHelper } from 'shared/utils/httpRequestHelper'
 
 // action creators and reducers in one file as suggested by https://github.com/erikras/ducks-modular-redux
 const REQUEST_DISCOVERY_SHEET = 'REQUEST_DISCOVERY_SHEET'
 const RECEIVE_DISCOVERY_SHEET = 'RECEIVE_DISCOVERY_SHEET'
-const REQUEST_SUCCESS_STORY = 'REQUEST_SUCCESS_STORY'
-const RECEIVE_SUCCESS_STORY = 'RECEIVE_SUCCESS_STORY'
-const REQUEST_ELASTICSEARCH_STATUS = 'REQUEST_ELASTICSEARCH_STATUS'
-const RECEIVE_ELASTICSEARCH_STATUS = 'RECEIVE_ELASTICSEARCH_STATUS'
-const REQUEST_MME = 'REQUEST_MME'
-const RECEIVE_MME = 'RECEIVE_MME'
 const REQUEST_SAMPLE_METADATA = 'REQUEST_SAMPLE_METADATA'
 const RECEIVE_SAMPLE_METADATA = 'RECEIVE_SAMPLE_METADATA'
-const RECEIVE_SAVED_VARIANT_TAGS = 'RECEIVE_SAVED_VARIANT_TAGS'
 const REQUEST_SEARCH_HASH_CONTEXT = 'REQUEST_SEARCH_HASH_CONTEXT'
 const RECEIVE_SEARCH_HASH_CONTEXT = 'RECEIVE_SEARCH_HASH_CONTEXT'
 const REQUEST_SEQR_STATS = 'REQUEST_SEQR_STATS'
 const RECEIVE_SEQR_STATS = 'RECEIVE_SEQR_STATS'
-const RECEIVE_PIPELINE_UPLOAD_STATS = 'RECEIVE_PIPELINE_UPLOAD_STATS'
-const UPDATE_STAFF_SAVED_VARIANT_TABLE_STATE = 'UPDATE_STAFF_VARIANT_STATE'
-const REQUEST_ALL_USERS = 'REQUEST_ALL_USERS'
-const RECEIVE_ALL_USERS = 'RECEIVE_ALL_USERS'
 
 
 // Data actions
-export const loadElasticsearchStatus = () => {
-  return (dispatch) => {
-    dispatch({ type: REQUEST_ELASTICSEARCH_STATUS })
-    new HttpRequestHelper('/api/staff/elasticsearch_status',
-      (responseJson) => {
-        dispatch({ type: RECEIVE_ELASTICSEARCH_STATUS, newValue: responseJson })
-      },
-      (e) => {
-        dispatch({ type: RECEIVE_ELASTICSEARCH_STATUS, error: e.message, newValue: { errors: [e.message] } })
-      },
-    ).get()
-  }
-}
-
-export const loadMme = () => {
-  return (dispatch) => {
-    dispatch({ type: REQUEST_MME })
-    new HttpRequestHelper('/api/staff/matchmaker',
-      (responseJson) => {
-        dispatch({ type: RECEIVE_DATA, updatesById: responseJson })
-        dispatch({ type: RECEIVE_MME, newValue: responseJson })
-      },
-      (e) => {
-        dispatch({ type: RECEIVE_MME, error: e.message, newValue: [] })
-      },
-    ).get()
-  }
-}
-
 const loadMultiProjectData = (requestAction, receiveAction, urlPath) => (projectGuid, filterValues) => {
   return (dispatch) => {
     if (projectGuid === 'all') {
@@ -121,42 +79,6 @@ export const loadDiscoverySheet = loadMultiProjectData(REQUEST_DISCOVERY_SHEET, 
 
 export const loadSampleMetadata = loadMultiProjectData(REQUEST_SAMPLE_METADATA, RECEIVE_SAMPLE_METADATA, 'sample_metadata')
 
-export const loadSuccessStory = (successStoryTypes) => {
-  return (dispatch) => {
-    if (successStoryTypes) {
-      dispatch({ type: REQUEST_SUCCESS_STORY })
-      new HttpRequestHelper(`/api/staff/success_story/${successStoryTypes}`,
-        (responseJson) => {
-          console.log(responseJson.errors)
-          dispatch({ type: RECEIVE_SUCCESS_STORY, newValue: responseJson.rows })
-        },
-        (e) => {
-          dispatch({ type: RECEIVE_SUCCESS_STORY, error: e.message, newValue: [] })
-        },
-      ).get()
-    }
-  }
-}
-
-export const loadAllUsers = () => {
-  return (dispatch, getState) => {
-    const { allUsers } = getState()
-    if (allUsers && allUsers.length) {
-      return
-    }
-
-    dispatch({ type: REQUEST_ALL_USERS })
-    new HttpRequestHelper('/api/staff/get_all_users',
-      (responseJson) => {
-        dispatch({ type: RECEIVE_ALL_USERS, newValue: responseJson })
-      },
-      (e) => {
-        dispatch({ type: RECEIVE_ALL_USERS, error: e.message, newValue: [] })
-      },
-    ).get()
-  }
-}
-
 export const loadSeqrStats = () => {
   return (dispatch) => {
     dispatch({ type: REQUEST_SEQR_STATS })
@@ -168,52 +90,6 @@ export const loadSeqrStats = () => {
         dispatch({ type: RECEIVE_SEQR_STATS, error: e.message, newValue: {} })
       },
     ).get()
-  }
-}
-
-export const uploadQcPipelineOutput = (values) => {
-  return (dispatch) => {
-    return new HttpRequestHelper('/api/staff/upload_qc_pipeline_output',
-      (responseJson) => {
-        dispatch({ type: RECEIVE_PIPELINE_UPLOAD_STATS, newValue: responseJson })
-      },
-      (e) => {
-        if (e.body && e.body.errors) {
-          throw new SubmissionError({ _error: e.body.errors })
-        } else {
-          throw new SubmissionError({ _error: [e.message] })
-        }
-      },
-    ).post(values)
-  }
-}
-
-export const loadSavedVariants = ({ tag, gene = '' }) => {
-  return (dispatch, getState) => {
-    // Do not load if already loaded
-    if (tag) {
-      if (getState().savedVariantTags[tag]) {
-        return
-      }
-    } else if (!gene) {
-      return
-    }
-
-    dispatch({ type: REQUEST_SAVED_VARIANTS })
-    new HttpRequestHelper(`/api/staff/saved_variants/${tag}`,
-      (responseJson) => {
-        if (tag && !gene) {
-          dispatch({
-            type: RECEIVE_SAVED_VARIANT_TAGS,
-            updates: { [tag]: true },
-          })
-        }
-        dispatch({ type: RECEIVE_DATA, updatesById: responseJson })
-      },
-      (e) => {
-        dispatch({ type: RECEIVE_DATA, error: e.message, updatesById: {} })
-      },
-    ).get({ gene })
   }
 }
 
@@ -264,7 +140,6 @@ export const loadProjectGroupContext = (projectCategoryGuid, addElementCallback)
     }
 
     else {
-      dispatch({ type: REQUEST_PROJECTS })
       new HttpRequestHelper('/api/search_context',
         (responseJson) => {
           dispatch({ type: RECEIVE_DATA, updatesById: responseJson })
@@ -278,34 +153,14 @@ export const loadProjectGroupContext = (projectCategoryGuid, addElementCallback)
   }
 }
 
-export const updateStaffSavedVariantTable = updates => ({ type: UPDATE_STAFF_SAVED_VARIANT_TABLE_STATE, updates })
-
 export const reducers = {
   sampleMetadataLoading: loadingReducer(REQUEST_SAMPLE_METADATA, RECEIVE_SAMPLE_METADATA),
   sampleMetadataRows: createSingleValueReducer(RECEIVE_SAMPLE_METADATA, []),
   discoverySheetLoading: loadingReducer(REQUEST_DISCOVERY_SHEET, RECEIVE_DISCOVERY_SHEET),
   discoverySheetRows: createSingleValueReducer(RECEIVE_DISCOVERY_SHEET, []),
-  successStoryLoading: loadingReducer(REQUEST_SUCCESS_STORY, RECEIVE_SUCCESS_STORY),
-  successStoryRows: createSingleValueReducer(RECEIVE_SUCCESS_STORY, []),
-  elasticsearchStatusLoading: loadingReducer(REQUEST_ELASTICSEARCH_STATUS, RECEIVE_ELASTICSEARCH_STATUS),
-  elasticsearchStatus: createSingleValueReducer(RECEIVE_ELASTICSEARCH_STATUS, {}),
-  mmeLoading: loadingReducer(REQUEST_MME, RECEIVE_MME),
-  mmeMetrics: createSingleValueReducer(RECEIVE_MME, {}, 'metrics'),
-  mmeSubmissions: createSingleValueReducer(RECEIVE_MME, [], 'submissions'),
-  projectGroupContextLoading: loadingReducer(REQUEST_PROJECTS, RECEIVE_DATA),
-  savedVariantTags: createSingleObjectReducer(RECEIVE_SAVED_VARIANT_TAGS),
   searchHashContextLoading: loadingReducer(REQUEST_SEARCH_HASH_CONTEXT, RECEIVE_SEARCH_HASH_CONTEXT),
   seqrStatsLoading: loadingReducer(REQUEST_SEQR_STATS, RECEIVE_SEQR_STATS),
   seqrStats: createSingleValueReducer(RECEIVE_SEQR_STATS, {}),
-  qcUploadStats: createSingleValueReducer(RECEIVE_PIPELINE_UPLOAD_STATS, {}),
-  staffSavedVariantTableState: createSingleObjectReducer(UPDATE_STAFF_SAVED_VARIANT_TABLE_STATE, {
-    categoryFilter: SHOW_ALL,
-    sort: SORT_BY_XPOS,
-    page: 1,
-    recordsPerPage: 25,
-  }, false),
-  allUsers: createSingleValueReducer(RECEIVE_ALL_USERS, [], 'users'),
-  allUsersLoading: loadingReducer(REQUEST_ALL_USERS, RECEIVE_ALL_USERS),
 }
 
 const rootReducer = combineReducers(reducers)
